Fix: Config Properties 구조 수정중
